refactor dashboard landing components
